Added InventoryToInstall to (update + save)Install
Added option to choose an inventory when making a new install to form a
new InventoryToInstall entry. Inventory to install is only available for
Real devices. We can
only choose between inventories that are currently not installed. If we
decide to remove a InventoryToInstall for selected Install, we can
simply use the x button and
press update.

Changes:

------------------------------------------------------------------------------------
idods.py:

------------------------------------------------------------------------------------
retrieveInventory optional parameter not_installed added. That enables
us now to return inventories that have not yet been installed.

retrieveInventoryTonstall all parameters are now optional, since a
retrieving method is supposed to be based on optional parameters and
because it was only useful if we
had all the parameters. But in that case we've already determined the
Install that we were looking for.

------------------------------------------------------------------------------------
dataapi/test_idods.py

------------------------------------------------------------------------------------
New tests were made due to changes idods.py, to make sure everything
that was changed works properly.

tests updated:
- testInventoryToInstall - added test for retrieving inventories that
are not installed.

------------------------------------------------------------------------------------
views.py

------------------------------------------------------------------------------------
Parameter not_installed added for retrieveInventory

------------------------------------------------------------------------------------
controllers.js

------------------------------------------------------------------------------------
Added $scope.inventories that holds all retrieved inventories.
If we are on an update form, we have to first retrieve an inventory that
is bound on that device via InventoryToInstall if such an inventory
exists.

------------------------------------------------------------------------------------
values.js

------------------------------------------------------------------------------------
Added not_installed to list of parameters in Install.

------------------------------------------------------------------------------------
install.html

------------------------------------------------------------------------------------
Added a <select> that shows all available inventories, together with an
X icon, that resets the <select> field.